[UPDATE] created a seperate context for wishlist, overall wishlist functionality improved
