+++ src/main.jsx
@@ -9,6 +9,7 @@ import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
 import { HelmetProvider } from "react-helmet-async";
 import AuthContextProvider from "./contexts/AuthContext.jsx";
 import { CartProvider } from "./contexts/CartContext.jsx";
+import { WishlistProvider } from "./contexts/WishlistContext.jsx";
 
 // Create a client
 const queryClient = new QueryClient();
@@ -18,9 +19,11 @@ createRoot(document.getElementById("root")).render(
     <QueryClientProvider client={queryClient}>
       <HelmetProvider>
         <CartProvider>
-          <AuthContextProvider>
-            <RouterProvider router={router} />
-          </AuthContextProvider>
+          <WishlistProvider>
+            <AuthContextProvider>
+              <RouterProvider router={router} />
+            </AuthContextProvider>
+          </WishlistProvider>
         </CartProvider>
       </HelmetProvider>
       <ReactQueryDevtools initialIsOpen={false} />
